test(agents): Start unit tests for agent refactor
