+++ .vscode/launch.json
@@ -13,8 +13,7 @@
                 "--timeout",
                 "999999",
                 "--colors",
-                "${workspaceFolder}\\**\\demo.test.ts"
-                // "${workspaceFolder}\\**\\*.test.ts"
+                "${workspaceFolder}\\**\\test\\unit\\**\\*.test.ts"
             ],
             "console": "integratedTerminal",
             "internalConsoleOptions": "neverOpen",
