+++ test/mocha.opts
@@ -1,4 +1,4 @@
 --compilers ts-node/register
 --require source-map-support/register
 --full-trace
-test/demo.test.ts
+test/unit/**/*.test.ts
